updates around symbols and delay initialization
remove use of from_c_ptr
remove second call to updt_params
unsigned annotation
delay instantation of smt_params in dl_context
cache xor_solver symbol
beef is unsigned too

+++ src/ast/seq_decl_plugin.cpp
@@ -523,7 +523,7 @@ void seq_decl_plugin::init() {
     if (m_init) return;
     ast_manager& m = *m_manager;
     m_init = true;
-    sort* A = m.mk_uninterpreted_sort(symbol((unsigned)0));
+    sort* A = m.mk_uninterpreted_sort(symbol(0u));
     sort* strT = m_string;
     parameter paramA(A);
     parameter paramS(strT);
